Trying to fix a Windows installation issue

+++ QuickOSM/quick_osm.py
@@ -85,6 +85,8 @@ class QuickOSMPlugin:
     @staticmethod
     def extract_zip_file():
         """ Extract or not the ZIP resources/i18n. """
+        # resources/i18n is created with plugin installation
+
         preset_translation_path = join(resources_path(), 'i18n')
         version_file_translation_path = join(resources_path(), 'i18n', 'version.txt')
         version_file_presets = join(resources_path(), 'JOSM_preset', 'version.txt')
@@ -92,6 +94,7 @@ class QuickOSMPlugin:
             # Legacy before 2.1.0
             shutil.rmtree(preset_translation_path)
             LOGGER.info('The version does not exist in the i18n folder, the folder needs to be unzipped.')
+
         if os.path.isdir(preset_translation_path) and os.path.isfile(version_file_translation_path):
             with open(version_file_translation_path, 'r', encoding='utf8') as check:
                 old_version = check.read().strip()
@@ -108,6 +111,7 @@ class QuickOSMPlugin:
                         old_version, new_version
                     )
                 )
+
         if not os.path.isdir(preset_translation_path):
             if os.path.isfile(preset_translation_path + '.zip'):
                 result = QgsZipUtils.unzip(preset_translation_path + '.zip', resources_path())
@@ -117,13 +121,22 @@ class QuickOSMPlugin:
                     LOGGER.info('Preset translations have been loaded and unzipped.')
                     files = os.listdir(preset_translation_path)
                     for file in files:
-                        file_path = join(preset_translation_path, file)
-                        if '-r' in file:
-                            new_file_path = join(preset_translation_path, file.replace('-r', '_'))
-                            os.rename(file_path, new_file_path)
-                        elif '-' in file:
-                            new_file_path = join(preset_translation_path, file.replace('-', '_'))
-                            os.rename(file_path, new_file_path)
+                        try:
+                            file_path = join(preset_translation_path, file)
+                            if '-r' in file:
+                                new_file_path = join(preset_translation_path, file.replace('-r', '_'))
+                                if not os.path.isfile(new_file_path):
+                                    # Ticket #418
+                                    os.rename(file_path, new_file_path)
+                            elif '-' in file:
+                                new_file_path = join(preset_translation_path, file.replace('-', '_'))
+                                if not os.path.isfile(new_file_path):
+                                    # Ticket #418
+                                    os.rename(file_path, new_file_path)
+                        except FileExistsError:
+                            LOGGER.critical(
+                                'Error about existing file when extracting the ZIP file about {}'.format(
+                                    file))
             else:
                 os.mkdir(preset_translation_path)
